+++ examples/mixed-ic-lua/supplier_logic.lua
@@ -1,0 +1,12 @@
+local supplier = {}
+
+local vendors = {
+    [-1301215609] = "iron-vendor",
+    [226410516] = "gold-vendor",
+}
+
+function supplier.vendor_for(requested_item)
+    return vendors[requested_item]
+end
+
+return supplier